Organize file structure and remove obsolete boilerplate files before starting to build the app.
