+++ src/components/App.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import '../styles/App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <h1>React Weather App</h1>
+      </header>
+    </div>
+  );
+}
+
+export default App;
